requirements.txt
psycopg[binary]==3.2.13
python-dotenv==1.0.1
feedparser==6.0.11
yfinance==1.2.0
vaderSentiment==3.3.2